feat(sdr): Phase 1 step 1 — utils/sdr_analysis.py, tested against REAL moc5 IQ
Pure analysis of one Airspy burst per design rev 2: absolute dBFS per-bin
floor; frame gates for our own near-field TX (any fleet channel above
-40 dBFS) and blockers (frame floor jump); class A off-channel carriers
with a measured per-(window, gain) spur map; class C 125 kHz foreign slices
tagged, not dropped, on our own IM3 products; class B per-bin floor delta;
>=2-run persistence. Statuses ok / overload / unjudgeable — never "clean".

Fixtures are 20 ms crops RECORDED on moc5 (not synthetic — the review
showed clean synthetic passes the analog failure trivially). The control
that can fail: with the own-TX gate disabled, the real near-field LF burst
is falsely reported as foreign energy (Phase 0's failure, reproduced from
recorded IQ); with it, the burst is `unjudgeable`.

Defect found and fixed in BOTH copies: np.abs(int16 -32768) overflows to
-32768, and the Airspy's int16 is (adc-2048)<<4 — an ADC pinned at the
negative rail read as unclipped (planted: old clip_frac 0.0, new 0.02).

Open-input reference (operator removed only the Airspy antenna): class B
is blind at gain 0-10 (antenna adds ~0.1 dB at 10, ~4.5 dB at 21) -> B at
high gain; 911.9985 is the Airspy's own spur; 937.515 external; near-field
LF uses half of full scale at gain 10 (~6 dB headroom).

Verified: 22 passed rc=0; regression guards 82 passed rc=0; lint rc=0.
Full suite not run (new module + tests + a one-line script fix).
Author-built, not yet non-author reviewed.

+++ scripts/sdr_fleet_channels.py
@@ -100,7 +100,7 @@ def capture(center_mhz: float, n: int, gain: int, tmpdir: str):
 
 def analyse(raw: np.ndarray, center_mhz: float, channels):
     """Per-channel stats for one burst. Pure."""
-    clip = float(np.mean(np.abs(raw) >= 0.95 * FULL_SCALE))
+    clip = float(np.mean(np.abs(raw.astype(np.int32)) >= 0.95 * FULL_SCALE))
     iq = raw[0::2].astype(np.float32) + 1j * raw[1::2].astype(np.float32)
     frames = iq.size // FFT
     spec = np.fft.fftshift(
